✨ feat: create  content query, contents query
